Move classes to com.getsixtyfour.openvpnmgmt.android
