+++ app/src/main/java/com/getsixtyfour/openvpnmgmt/android/VPNAuthenticationHandler.java
@@ -1,4 +1,4 @@
-package com.getsixtyfour.openvpnmgmt.android.core;
+package com.getsixtyfour.openvpnmgmt.android;
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -6,7 +6,6 @@ import android.content.SharedPreferences;
 import androidx.annotation.NonNull;
 import androidx.preference.PreferenceManager;
 
-import com.getsixtyfour.openvpnmgmt.android.Constants;
 import com.getsixtyfour.openvpnmgmt.net.UsernamePasswordHandler;
 
 /**
